refactor: slimmer SummaryCard (823 lines down to 599)
New files:
* `hideable_container.dart`: used to be litterally in `summary_card.dart`
* `product_questions_widget.dart`: used to be as a method in `summary_card.dart`

Impacted files:
* `product_cards_helper.dart`: moved there 2 methods from `summary_card.dart`
* `scan_product_card.dart`: minor refactoring
* `summary_card.dart`: refactored moving code to other files and marginally setting fields and methods as private
